Enum gettext::Error

pub enum Error {
    BadMagic,
    DecodingError,
    Eof,
    Io(Error),
    MalformedMetadata,
    MisplacedMetadata,
    UnknownEncoding,
}

Represents an error encountered while parsing an MO file.

Variants

BadMagic

An incorrect magic number has been encountered

DecodingError

An invalid byte sequence for the given encoding has been encountered

Eof

An unexpected EOF occured

Io

An I/O error occured

MalformedMetadata

Incorrect syntax encountered while parsing the meta information

MisplacedMetadata

Meta information string was not the first string in the catalog

UnknownEncoding

An unknown encoding was specified in the metadata