Struct gettext::ParseOptions

pub struct ParseOptions {
    // some fields omitted
}

ParseOptions allows setting options for parsing MO catalogs.

Examples

use std::fs::File;

extern crate encoding;
use encoding::all::ISO_8859_1;

let file = File::open("french.mo").unwrap();
let catalog = ParseOptions::new().force_encoding(ISO_8859_1).parse(file).unwrap();

Methods

impl ParseOptions

fn new() -> Self

Returns a new instance of ParseOptions with default options.

fn parse<R: Read>(self, reader: R) -> Result<Catalog, Error>

Tries to parse the catalog from the given reader using the specified options.

fn force_encoding(self, encoding: EncodingRef) -> Self

Forces a use of a specific encoding when parsing strings from a catalog. If this option is not enabled, the parser tries to use the encoding specified in the metadata or UTF-8 if metadata is non-existent.

fn force_plural<T: Fn(u64) -> usize + 'static>(self, plural: T) -> Self

Forces a use of the given plural formula for deciding the proper plural form for a message. If this option is not enabled, the parser uses the default formula (n != 1).